Iterate over ascending fs::path
