+++ src/pf/existing/detect_project_root.cpp
@@ -5,6 +5,8 @@
 #include <string>
 #include <utility>
 
+#include <boost/range/algorithm/find_if.hpp>
+
 namespace fs = pf::fs;
 
 namespace {
@@ -26,16 +28,14 @@ boost::optional<fs::path> parse_cmakecache_homedir(fs::path const& cmakecache) {
 }
 }  // namespace
 
-boost::optional<fs::path> pf::detect_project_root(fs::path cur_dir) {
-    for (; cur_dir.has_parent_path(); cur_dir = cur_dir.parent_path()) {
+boost::optional<fs::path> pf::detect_project_root(fs::path from_dir) {
+    for (auto const& cur_dir : pf::ascending_range(from_dir)) {
         if (fs::exists(cur_dir / "CMakeLists.txt")) {
             // If we see a CMakeLists.txt, but the parent directory also has a CMakeLists.txt,
             // assume that we are a child of that parent directory
-            for (fs::path parent = cur_dir.parent_path(); fs::exists(parent / "CMakeLists.txt");
-                 cur_dir         = std::exchange(parent, parent.parent_path())) {
-            }
-
-            return cur_dir;
+            return *boost::range::find_if(pf::ascending_range(cur_dir), [](auto const& path) {
+                return !fs::exists(path.parent_path() / "CMakeLists.txt");
+            });
         }
 
         fs::path const maybe_cmakecache = cur_dir / "CMakeCache.txt";
