+++ src/pf/fs/ascending_iterator.hpp
@@ -1,0 +1,54 @@
+#ifndef PF_FS_ASCENDING_ITERATOR_HPP_INCLUDED
+#define PF_FS_ASCENDING_ITERATOR_HPP_INCLUDED
+
+#include <iterator>
+#include <utility>
+
+#include <pf/fs/core.hpp>
+
+namespace pf {
+
+class ascending_iterator {
+public:
+    using difference_type   = std::ptrdiff_t;
+    using value_type        = fs::path;
+    using pointer           = value_type const*;
+    using reference         = value_type const&;
+    using iterator_category = std::input_iterator_tag;
+
+    ascending_iterator() = default;
+
+    explicit ascending_iterator(fs::path p)
+        : current_{std::move(p)} {}
+
+    ascending_iterator& operator++() {
+        current_ = current_.parent_path();
+        return *this;
+    }
+
+    ascending_iterator operator++(int) {
+        auto cpy = *this;
+
+        ++*this;
+
+        return cpy;
+    }
+
+    reference operator*() const { return current_; }
+    pointer   operator->() const { return &current_; }
+
+    friend bool operator==(ascending_iterator const& lhs, ascending_iterator const& rhs) {
+        return lhs.current_ == rhs.current_;
+    }
+
+    friend bool operator!=(ascending_iterator const& lhs, ascending_iterator const& rhs) {
+        return !(lhs == rhs);
+    }
+
+private:
+    fs::path current_ = {};
+};
+
+}  // namespace pf
+
+#endif  // PF_FS_ASCENDING_ITERATOR_HPP_INCLUDED
